Store longitudinal speed in DB
Also:

* Introduce a `ghci` file for common repl options
* Move the Database module to Ephemeris internal
* Re-populate the ephemeris DB
